@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"sync"
 
 	"github.com/hashicorp/go-multierror"
 
@@ -12,7 +11,12 @@ import (
 	"github.com/libp2p/go-libp2p/core/protocol"
 )
 
-func (r *Replica) send(to peer.ID, msg interface{}, protocol protocol.ID) error {
+func (r *Replica) send(ctx context.Context, to peer.ID, msg any, protocol protocol.ID) error {
+
+	ctx, span := r.tracer.Start(ctx, msgSendSpanName(msg, spanMessageSend))
+	defer span.End()
+
+	saveTraceContext(ctx, msg)
 
 	// Serialize the message.
 	payload, err := json.Marshal(msg)
@@ -21,7 +25,7 @@ func (r *Replica) send(to peer.ID, msg interface{}, protocol protocol.ID) error 
 	}
 
 	// We don't want to wait indefinitely.
-	ctx, cancel := context.WithTimeout(context.Background(), r.cfg.NetworkTimeout)
+	ctx, cancel := context.WithTimeout(ctx, r.cfg.NetworkTimeout)
 	defer cancel()
 
 	// Send message.
@@ -34,7 +38,12 @@ func (r *Replica) send(to peer.ID, msg interface{}, protocol protocol.ID) error 
 }
 
 // broadcast sends message to all peers in the replica set.
-func (r *Replica) broadcast(msg interface{}) error {
+func (r *Replica) broadcast(ctx context.Context, msg any) error {
+
+	ctx, span := r.tracer.Start(ctx, msgSendSpanName(msg, spanMessageBroadcast))
+	defer span.End()
+
+	saveTraceContext(ctx, msg)
 
 	// Serialize the message.
 	payload, err := json.Marshal(msg)
@@ -45,48 +54,37 @@ func (r *Replica) broadcast(msg interface{}) error {
 	ctx, cancel := context.WithTimeout(context.Background(), r.cfg.NetworkTimeout)
 	defer cancel()
 
-	var (
-		wg       sync.WaitGroup
-		multierr *multierror.Error
-		lock     sync.Mutex
-	)
-
+	var errGroup multierror.Group
 	for _, target := range r.peers {
+		target := target
+
 		// Skip self.
 		if target == r.id {
 			continue
 		}
 
-		wg.Add(1)
-
 		// Send concurrently to everyone.
-		go func(peer peer.ID) {
-			defer wg.Done()
+		errGroup.Go(func() error {
 
 			// NOTE: We could potentially retry sending if we fail once. On the other hand, somewhat unlikely they're
 			// back online split second later.
-
-			err := r.host.SendMessageOnProtocol(ctx, peer, payload, r.protocolID)
+			err := r.host.SendMessageOnProtocol(ctx, target, payload, r.protocolID)
 			if err != nil {
-
-				lock.Lock()
-				defer lock.Unlock()
-
-				multierr = multierror.Append(multierr, err)
+				return fmt.Errorf("peer send error (peer: %v): %w", target.String(), err)
 			}
-		}(target)
+
+			return nil
+		})
 	}
 
-	wg.Wait()
-
 	// If all went well, just return.
-	sendErr := multierr.ErrorOrNil()
-	if sendErr == nil {
+	sendErr := errGroup.Wait()
+	if sendErr.ErrorOrNil() == nil {
 		return nil
 	}
 
 	// Warn if we had more send errors than we bargained for.
-	errCount := uint(multierr.Len())
+	errCount := uint(sendErr.Len())
 	if errCount > r.f {
 		r.log.Warn().Uint("f", r.f).Uint("errors", errCount).Msg("broadcast error count higher than pBFT f value")
 	}
