@@ -5,15 +5,19 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/armon/go-metrics"
 	"github.com/asaskevich/govalidator"
-	"github.com/rs/zerolog"
-
 	"github.com/libp2p/go-libp2p"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	"github.com/libp2p/go-libp2p/core/crypto"
 	"github.com/libp2p/go-libp2p/core/host"
+	"github.com/libp2p/go-libp2p/core/peer"
 	rcmgr "github.com/libp2p/go-libp2p/p2p/host/resource-manager"
+	"github.com/libp2p/go-libp2p/p2p/net/connmgr"
+	"github.com/libp2p/go-libp2p/p2p/protocol/circuitv2/relay"
+	webrtc "github.com/libp2p/go-libp2p/p2p/transport/webrtc"
 	ma "github.com/multiformats/go-multiaddr"
+	"github.com/rs/zerolog"
 )
 
 // Host represents a new libp2p host.
@@ -23,7 +27,8 @@ type Host struct {
 	log zerolog.Logger
 	cfg Config
 
-	pubsub *pubsub.PubSub
+	pubsub  *pubsub.PubSub
+	metrics *metrics.Metrics
 }
 
 // New creates a new Host.
@@ -53,6 +58,7 @@ func New(log zerolog.Logger, address string, port uint, options ...func(*Config)
 	opts := []libp2p.Option{
 		libp2p.ListenAddrStrings(addresses...),
 		libp2p.DefaultTransports,
+		libp2p.Transport(webrtc.New),
 		libp2p.DefaultMuxers,
 		libp2p.DefaultSecurity,
 		libp2p.NATPortMap(),
@@ -118,15 +124,47 @@ func New(log zerolog.Logger, address string, port uint, options ...func(*Config)
 		opts = append(opts, addrFactory)
 	}
 
+	if cfg.ConnectionLimit > 0 {
+
+		lo := cfg.ConnectionLimit
+		hi := float64(cfg.ConnectionLimit) * 1.2
+
+		cm, err := connmgr.NewConnManager(int(lo), int(hi))
+		if err != nil {
+			return nil, fmt.Errorf("could not create connection manager (lo: %v, hi: %v)", lo, hi)
+		}
+
+		// Protect boot nodes from pruning.
+		for _, addr := range cfg.BootNodes {
+			info, err := peer.AddrInfoFromP2pAddr(addr)
+			if err != nil {
+				continue
+			}
+
+			cm.Protect(info.ID, "boot-node")
+		}
+
+		opts = append(opts, libp2p.ConnectionManager(cm))
+	}
+
 	// Create libp2p host.
 	h, err := libp2p.New(opts...)
 	if err != nil {
 		return nil, fmt.Errorf("could not create libp2p host: %w", err)
 	}
 
+	if cfg.EnableP2PRelay {
+		log.Info().Msg("enabling p2p relay...")
+		_, err = relay.New(h)
+		if err != nil {
+			return nil, fmt.Errorf("could not create relay: %w", err)
+		}
+	}
+
 	host := Host{
-		log: log.With().Str("component", "host").Logger(),
-		cfg: cfg,
+		log:     log,
+		cfg:     cfg,
+		metrics: metrics.Default(),
 	}
 	host.Host = h
 
